Create internal 'liveMode' for XVIZStreamBuffer
This mode is detected when metadata is missing the log_info.start_time.
This effects how the initial request for data is made and internal
tracking in the stream buffer.

For now this will make an assumption on the buffer length being the size
passed in via bufferLength, or a default of 30 seconds.