Add URL passing to the XVIZ Loaders
Rather than construct with fix parameters, allow the getting-started
to use the URL determine which data is loaded.

Default behavior w/o a URL path is the same, but if a path is provided
the stream and live loaders will construct their request using the url
path and params

Supports loading multiple logs easily and is useful for testing.